+++ acceptance/pipelines/stop/output.txt
@@ -10,13 +10,8 @@ View your pipeline my_pipeline here: [DATABRICKS_URL]/pipelines/[UUID]?o=[NUMID]
 Update URL: [DATABRICKS_URL]/#joblist/pipelines/[UUID]/updates/[UUID]
 
 Update ID: [UUID]
-Warn: Failed to fetch events for update [UUID]: failed to get next event: The specified pipeline events was not found.
-State: COMPLETED
-Creation Time: 1970-01-01T00:00:00Z
-Full Refresh: NO
-Validate Only: NO
-Pipeline Spec: (No config available)
-
+Update [UUID] for pipeline completed successfully.
+Refresh: default refresh-all
 
 === Stopping pipeline
 >>> [PIPELINES] stop
